+++ src/tests/perception_module_test/IdentifyAgentsCommandTest.cpp
@@ -13,7 +13,7 @@
 
 BOOST_FIXTURE_TEST_CASE( identify_all_agents_in_zone_command_notifies_perceptions, sword::perception::ModuleFixture )
 {
-    const size_t identifier = 3;
+    const unsigned int identifier = 3;
     const SWORD_Model* other = core::Convert( &model[ "entities/other" ] );
     model[ "entities/other/pion" ] = 43;
     core::Model& zone = model[ "entities" ][ identifier ][ "perceptions/notifications/agents-in-zone" ];
